Develop working

+++ src/main/java/acme/forms/EpicureDashboard.java
@@ -15,9 +15,9 @@ public class EpicureDashboard implements Serializable{
 
 	private static final long serialVersionUID = 1L;
 	
-	Map<String, Integer> totalNDishesOfState;
+	Map<String, Long> totalNDishesOfState;
 
-  Map<Pair<String, String>,Double> averageBudgetFineDishesOfStateByCurrency;
+	Map<Pair<String, String>,Double> averageBudgetFineDishesOfStateByCurrency;
 	Map<Pair<String, String>,Double> deviationBudgetFineDishesOfStateByCurrency;
 	Map<Pair<String, String>,Double> minBudgetFineDishesOfStateByCurrency;
 	Map<Pair<String, String>,Double> maxBudgetFineDishesOfStateByCurrency;
